Issues/128 code review

+++ BostonFreedomTrail/Utils/Constants.swift
@@ -30,11 +30,11 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 import Foundation
 
-enum SegueConstants : String {
+enum SegueConstants: String {
     case MapToPlacemarkSegueIdentifier = "MapToPlacemarkSegueIdentifier"
 }
 
-enum TabBarControllerTags : NSInteger {
+enum TabBarControllerTags: Int {
     case MapViewTag = 0
     case VirtualTourViewTag = 1
     case AboutViewTag = 2
